animal_vision_app/requirements.txt
Flask==3.0.0
tensorflow
pillow
numpy
